refactor: migrate double→BigDecimal for all prices, String→LocalDateTime for order date
- Rewrite schema.sql/drop.sql to match JPA entity @Table names
- Fix column types: double→decimal(10,2), varchar→timestamp
- Add missing FK constraints, indexes, and join tables
- Migrate 7 entities: Cart, CartProduct, Order, OrderProduct, Product,
  ProductPosterSize, LocalizedPrice
- Update PostersProperties, SessionService, PriceFormatter,
  DataImportService, CartController, CheckoutController,
  CatalogController for BigDecimal compatibility
- Change Order.orderDate from String to LocalDateTime

+++ src/main/java/com/xceptance/posters/util/PriceFormatter.java
@@ -1,5 +1,6 @@
 package com.xceptance.posters.util;
 
+import java.math.BigDecimal;
 import java.text.NumberFormat;
 import java.util.Locale;
 import java.util.Map;
@@ -47,7 +48,7 @@ public final class PriceFormatter
      * @param urlLocale the URL locale string (e.g. "de-DE", "en-GB", "sv-SE")
      * @return formatted price string (e.g. "17,00 €", "185,30 kr")
      */
-    public static String format(double price, String urlLocale)
+    public static String format(BigDecimal price, String urlLocale)
     {
         Locale javaLocale = toJavaLocale(urlLocale);
         NumberFormat fmt = NumberFormat.getCurrencyInstance(javaLocale);
